consolidate nfs server rules into one class and enforce
