+++ data/cis/cis_AlmaLinux_9_rules.yaml
@@ -93,7 +93,7 @@ cis_security_hardening::benchmark::almalinux::v9:
         - telnet_server
         - dnsmasq
         - mta_local
-        - nfs_utils
+        - nfs
         - rpcbind
         - rsyncd
     service_clients:
