+++ data/cis/cis_CentOS_7_params.yaml
@@ -141,7 +141,7 @@ cis_security_hardening::rules::avahi::enforce: true
 cis_security_hardening::rules::cups::enforce: true
 cis_security_hardening::rules::dhcp::enforce: true
 cis_security_hardening::rules::ldapd::enforce: true
-cis_security_hardening::rules::nfs_utils::enforce: true
+cis_security_hardening::rules::nfs::enforce: true
 cis_security_hardening::rules::rpcbind::enforce: true
 cis_security_hardening::rules::bind::enforce: true
 cis_security_hardening::rules::dnsmasq::enforce: true
