+++ data/cis/cis_RedHat_7_rules.yaml
@@ -100,7 +100,7 @@ cis_security_hardening::benchmark::redhat::v7:
         - samba
         - vsftp
         - dovecot
-        - nfs_utils
+        - nfs
         - nis
         - cups
         - rpcbind
