+++ data/cis/cis_RedHat_9_params.yaml
@@ -184,12 +184,11 @@ cis_security_hardening::rules::net_snmp::enforce: true
 cis_security_hardening::rules::mta_local::enforce: true
 cis_security_hardening::rules::rsyncd::enforce: true
 cis_security_hardening::rules::nis::enforce: true
-cis_security_hardening::rules::nfs: true
 cis_security_hardening::rules::tftp_server::enforce: true
 cis_security_hardening::rules::telnet_server::enforce: true
 cis_security_hardening::rules::dnsmasq::enforce: true
-cis_security_hardening::rules::nfs_utils::enforce: true
-cis_security_hardening::rules::nfs_utils::uninstall: true
+cis_security_hardening::rules::nfs::enforce: true
+cis_security_hardening::rules::nfs::uninstall: true
 cis_security_hardening::rules::rngd::enforce: true
 cis_security_hardening::rules::abrt::enforce: true
 cis_security_hardening::rules::sendmail::enforce: true
