+++ data/cis/cis_SLES_12_rules.yaml
@@ -71,7 +71,7 @@ cis_security_hardening::benchmark::sles::v12:
         - cups
         - dhcp
         - ldapd
-        - nfs_utils
+        - nfs
         - rpcbind
         - bind
         - vsftp
