+++ data/cis/cis_Ubuntu_22.04_params.yaml
@@ -129,7 +129,7 @@ cis_security_hardening::rules::avahi::enforce: true
 cis_security_hardening::rules::cups::enforce: true
 cis_security_hardening::rules::dhcp::enforce: true
 cis_security_hardening::rules::ldapd::enforce: true
-cis_security_hardening::rules::nfs_utils::enforce: true
+cis_security_hardening::rules::nfs::enforce: true
 cis_security_hardening::rules::bind::enforce: true
 cis_security_hardening::rules::vsftp::enforce: true
 cis_security_hardening::rules::httpd::enforce: true
